refactor(I/O): replace mio with custom `Mmap`

+++ include/zenkit/Stream.hh
@@ -222,7 +222,7 @@ namespace zenkit {
 
 			auto len = static_cast<ssize_t>(w->tell()) - size_off;
 			w->seek(size_off, Whence::BEG);
-			w->write_uint(len - sizeof(uint32_t));
+			w->write_uint(static_cast<uint32_t>(len) - sizeof(uint32_t));
 			w->seek(len, Whence::CUR);
 		}
 	} // namespace proto
